+++ src/main/java/org/la4j/factory/CompressedFactory.java
@@ -25,11 +25,12 @@ package org.la4j.factory;
 import java.util.Arrays;
 import java.util.Random;
 
+import org.la4j.matrix.Matrix;
 import org.la4j.vector.Vector;
 import org.la4j.vector.source.VectorSource;
 import org.la4j.vector.sparse.CompressedVector;
 
-public abstract class CompressedFactory implements Factory {
+public abstract class CompressedFactory extends Factory {
 
     private static final long serialVersionUID = 4071505L;
 
@@ -75,9 +76,7 @@ public abstract class CompressedFactory implements Factory {
     }
 
     @Override
-    public Vector createRandomVector(int length) {
-
-        Random random = new Random();
+    public Vector createRandomVector(int length, Random random) {
 
         int cardinality = length / DENSITY;